文件上传页面-修改接口量级返回值

See merge request pd_str/marketing!1765

+++ marketing-cs/src/main/java/com/br/marketing/service/Impl/LocalFileServiceImpl.java
@@ -64,7 +64,7 @@ public class LocalFileServiceImpl implements LocalFileService {
         return PageResultReturn.setPageResult(localFileList, current, pageSize);
     }
     @Override
-    public Integer allCount(String search, String apiCode, String uploadStartTime, String uploadEndTime, String fileType) {
+    public Long allCount(String search, String apiCode, String uploadStartTime, String uploadEndTime, String fileType) {
         if(StringUtils.isNotBlank(fileType)){
             fileType = "'"+fileType.replace(",", "','")+"'";
         }
